# backend/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RefuLearn API",
    "description": "A comprehensive API for the RefuLearn platform supporting refugee education and skill development",
    "version": "1.0.0",
    "contact": {
      "name": "RefuLearn Team",
      "email": "support@refulearn.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.refulearn.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    },
  "paths": {
...
    "/api/jobs": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get all jobs",
        "description": "Retrieve a list of job postings with optional filtering",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "location", "in": "query", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["full-time", "part-time", "contract", "internship"] } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Jobs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobs": { "type": "array", "items": { "$ref": "#/components/schemas/Job" } },
                        "pagination": { "type": "object" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["jobs"],
        "summary": "Create a job",
        "description": "Create a new job posting (employer only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "location": { "type": "string" },
                  "job_type": { "type": "string", "enum": ["Full Time", "Part Time", "Contract", "Internship"] },
                  "required_skills": { "type": "array", "items": { "type": "string" } },
                  "salary_range": { "type": "string" },
                  "application_deadline": { "type": "string", "format": "date-time" },
                  "is_active": { "type": "boolean" },
                  "remote_work": { "type": "boolean" }
                },
                "required": [
                  "title", "description", "location", "job_type", "required_skills", "salary_range", "application_deadline", "is_active", "remote_work"
                ]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Job created successfully" },
          "400": { "description": "Bad request" }
        }
      }
    },
    "/api/jobs/{jobId}": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get job by ID",
        "description": "Retrieve a job posting by its ID",
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Job retrieved successfully" },
          "404": { "description": "Job not found" }
        }
      },
      "put": {
        "tags": ["jobs"],
        "summary": "Update job",
        "description": "Update a job posting (employer/admin only). Only the employer who created the job (job.employer == user._id) or an admin can update the job.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "location": { "type": "string" },
                  "job_type": { "type": "string", "enum": ["Full Time", "Part Time", "Contract", "Internship"] },
                  "required_skills": { "type": "array", "items": { "type": "string" } },
                  "salary_range": { "type": "string" },
                  "application_deadline": { "type": "string", "format": "date-time" },
                  "is_active": { "type": "boolean" },
                  "remote_work": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Job updated successfully" },
          "400": { "description": "Bad request" },
          "404": { "description": "Job not found" },
          "403": { "description": "Not authorized to update this job" }
        }
      },
      "delete": {
        "tags": ["jobs"],
        "summary": "Delete job",
        "description": "Delete a job posting (employer/admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Job deleted successfully" },
          "404": { "description": "Job not found" }
        }
      }
    },
    "/api/jobs/employer/jobs": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get employer's jobs",
        "description": "Retrieve all jobs posted by the authenticated employer",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Jobs retrieved successfully" }
        }
      }
    },
    "/api/jobs/{jobId}/apply": {
      "post": {
        "tags": ["jobs"],
        "summary": "Apply for job",
        "description": "Apply for a job posting. Upload cover letter and resume as document files (PDF, DOC, DOCX, TXT). Expected salary is optional. Available to refugees, users, employers, and admins.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "coverLetter": { 
                    "type": "string", 
                    "format": "binary",
                    "description": "Cover letter document file (PDF, DOC, DOCX, TXT)"
                  },
                  "resume": { 
                    "type": "string", 
                    "format": "binary",
                    "description": "Resume document file (PDF, DOC, DOCX, TXT)"
                  },
                  "expectedSalary": { 
                    "type": "number",
                    "description": "Expected salary (optional)"
                  }
                },
                "required": ["coverLetter", "resume"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Application submitted successfully" },
          "400": { "description": "Bad request" },
          "404": { "description": "Job not found" }
        }
      }
    },
    "/api/jobs/applications/user": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get all job applications",
        "description": "Retrieve all job applications submitted by all users on the platform. Available to employers and admins only. Refugees and regular users cannot access this endpoint.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "All applications retrieved successfully" },
          "403": { "description": "Forbidden - Only employers and admins can access this endpoint" }
        }
      }
    },
    "/api/jobs/{jobId}/applications": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get job applications",
        "description": "Retrieve all applications for a specific job. Available to all employers and admins (not restricted to job owner).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Applications retrieved successfully" },
          "404": { "description": "Job not found" },
          "403": { "description": "Forbidden - Only employers and admins can access this endpoint" }
        }
      }
    },
    "/api/jobs/{jobId}/applications/{applicationId}": {
      "put": {
        "tags": ["jobs"],
        "summary": "Update application status",
        "description": "Update the status of a job application using the dropdown status selector. Available to all employers and admins (not restricted to job owner).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "applicationId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": { 
                    "type": "string", 
                    "enum": ["pending", "reviewed", "shortlisted", "rejected", "hired", "closed"],
                    "description": "Select application status",
                    "example": "reviewed"
                  },
                  "notes": { 
                    "type": "string",
                    "description": "Optional notes about the status change",
                    "example": "Candidate looks promising"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Application status updated successfully" },
          "404": { "description": "Application not found" },
          "403": { "description": "Forbidden - Only employers and admins can access this endpoint" }
        }
      }
    },
    "/api/jobs/{jobId}/analytics": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get job analytics",
        "description": "Retrieve analytics for a specific job. Available to all employers and admins (not restricted to job owner).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Analytics retrieved successfully" },
          "404": { "description": "Job not found" },
          "403": { "description": "Not authorized" }
        }
      }
    },
...
    "/api/admin/jobs": {
      "get": {
        "tags": ["admin"],
        "summary": "Get all jobs",
        "description": "Retrieve a list of all jobs with pagination and filtering",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Jobs retrieved successfully" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/admin/jobs/{jobId}": {
      "put": {
        "tags": ["admin"],
        "summary": "Update job",
        "description": "Update a job's information",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "status": { "type": "string" },
                  "employer": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Job updated successfully" },
          "404": { "description": "Job not found" }
        }
      }
    },
...
  }
}